+++ lightning/src/ln/mod.rs
@@ -23,6 +23,7 @@
 pub mod functional_test_utils;
 
 pub mod channelmanager;
+pub mod inbound_payment;
 pub mod msgs;
 pub mod peer_handler;
 pub mod chan_utils;
